+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.31.1-0 / Unreleased
+  * [Fixed] #951 Hover issue in unselectable radio
 
 # 2.31.0 / 2019-05-30
   * New extension **ext-grid** (experimental)<br>